backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
web3>=6.0
PyJWT>=2.8
bcrypt>=4.1
python-dotenv>=1.0
psycopg2-binary>=2.9
gunicorn>=21.2
